enhance: [2.6] upgrade knowhere to v2.6.12 and dependencies for C++20 compatibility
Cherry-pick master PR #47810 dependency upgrades plus knowhere v2.6.12
and CUDA 12.9.1 for GPU images.

Key dependency upgrades:
- opentelemetry-cpp: 1.8.1.1 -> 1.23.0
- protobuf: 3.21.4 -> 5.27.0
- grpc: 1.50.1 -> 1.67.1
- abseil: 20230125.3 -> 20250127.0
- folly: 2023.10.30.08 -> 2024.08.12.00
- fmt: 9.1.0 -> 11.0.2
- glog: 0.6.0 -> 0.7.1
- google-cloud-cpp: 2.5.0 -> 2.28.0

2.6-specific adaptations:
- Keep knowhere at v2.6.12 (not master's version)
- Keep milvus-storage at e159050, patch s3_fs.cpp for google-cloud-cpp 2.28
- Add gflags include to ConfigKnowhere.cpp for glog 0.7
- Keep s2n/1.6.0 override for OpenSSL 3.x FIPS detection

+++ internal/core/src/exec/expression/GISFunctionFilterExpr.cpp
@@ -249,7 +249,7 @@ PhyGISFunctionFilterExpr::EvalForDataSegment() {
         default: {
             ThrowInfo(NotImplemented,
                       "internal error: unknown GIS op : {}",
-                      expr_->op_);
+                      static_cast<int>(expr_->op_));
         }
     }
     return res_vec;
@@ -359,7 +359,9 @@ PhyGISFunctionFilterExpr::EvalForIndexSegment() {
                 // Distance-based operation - no prepared version
                 return left.dwithin(query_geometry, expr_->distance_);
             default:
-                ThrowInfo(NotImplemented, "unknown GIS op : {}", expr_->op_);
+                ThrowInfo(NotImplemented,
+                          "unknown GIS op : {}",
+                          static_cast<int>(expr_->op_));
         }
     };
